Relocate `SingleModuleGraph` to `turbopack-core`
While many use case-specific graphs belong in Next.js, this primitive should be kept in Turbopack as it will be used for other core usecases like chunking.

Test Plan: CI
