+++ crates/next-api/src/server_actions.rs
@@ -30,6 +30,7 @@ use turbopack_core::{
     context::AssetContext,
     file_source::FileSource,
     module::Module,
+    module_graph::SingleModuleGraph,
     output::OutputAsset,
     reference_type::{EcmaScriptModulesReferenceSubType, ReferenceType},
     resolve::ModulePart,
@@ -40,8 +41,6 @@ use turbopack_ecmascript::{
     chunk::EcmascriptChunkPlaceable, parse::ParseResult,
     tree_shake::asset::EcmascriptModulePartAsset, EcmascriptParsable,
 };
-
-use crate::module_graph::SingleModuleGraph;
 
 #[turbo_tasks::value]
 pub(crate) struct ServerActionsManifest {
