+++ turbopack/crates/turbopack-core/Cargo.toml
@@ -22,6 +22,7 @@ indexmap = { workspace = true }
 lazy_static = { workspace = true }
 once_cell = { workspace = true }
 patricia_tree = "0.5.5"
+petgraph = { workspace = true }
 ref-cast = "1.0.20"
 regex = { workspace = true }
 serde = { workspace = true, features = ["rc"] }
